Adiciona filtros na visão desktop do QMR
Filtro por nome, estado e partido disponíveis. Falta ajuste no CSS.
