+++ src/views/desktop-view/containers/deputados/DeputadosContainer.js
@@ -6,6 +6,11 @@ import infoPartidos from '../../../../data/partidos.json';
 import Card from 'material-ui/Card';
 //import TabsContainer from './TabsContainer.js';
 import FlipMove from 'react-flip-move';
+import { MenuItem } from 'material-ui/Menu';
+import Input, { InputLabel } from 'material-ui/Input';
+import { FormControl, FormHelperText } from 'material-ui/Form';
+import Select from 'material-ui/Select';
+import TextField from 'material-ui/TextField';
 import nomesVotacoes from '../../../../data/nomes_votacoes.json';
 
 const cardStyle = {
@@ -27,7 +32,8 @@ class DeputadosContainer extends Component {
     //this.nVotacoesPartido = this.calculaNVotacoesPartidos(this.getVotacoesPartidos());
     this.scoreDeputados = {};
     //this.scorePartidos = {};
-    this.state = { deputados: [] };
+    this.state = { deputados: [], deputadosPorNome: [], deputadosPorUf: [], deputadosPorPartido: [],
+      filterName: '', filterUf: 'TODOS', filterPartido: 'TODOS' };
   }
 
   componentWillReceiveProps(nextProps){
@@ -48,14 +54,44 @@ class DeputadosContainer extends Component {
   }
 
   render(){
+    let byName = this.state.filterName !== '' ? this.state.deputados.filter(d => this.state.deputadosPorNome.indexOf(d.key) !== -1) : this.state.deputados;
+    let byUf = this.state.filterUf !== 'TODOS' ? byName.filter(d => this.state.deputadosPorUf.indexOf(d.key) !== -1) : byName;
+    var deputadosAExibir = this.state.filterPartido !== 'TODOS' ? byUf.filter(d => this.state.deputadosPorPartido.indexOf(d.key) !== -1) : byUf;
 
     //console.log(this.props.scoreDeputados);
 
     return (
       <div>
+        <TextField label="Nome do Deputado" onChange={this.buscaNome.bind(this)} style={{width:'45%'}}/>
+        <FormControl>
+          <InputLabel htmlFor="uf-simple">UF</InputLabel>
+          <Select
+            value={this.state.filterUf}
+            onChange={this.buscaUF.bind(this)}
+            inputProps={{
+              name: 'uf',
+              id: 'uf-simple',
+            }}
+          >
+          {this.estados()}
+          </Select>
+        </FormControl>
+        <FormControl>
+          <InputLabel htmlFor="partido-simple">Partido</InputLabel>
+          <Select
+            value={this.state.filterPartido}
+            onChange={this.buscaPartido.bind(this)}
+            inputProps={{
+              name: 'partido',
+              id: 'partido-simple',
+            }}
+          >
+          {this.partidos()}
+          </Select>
+        </FormControl>
         <Card style={cardStyle}>
           <FlipMove>
-            {this.state.deputados.slice(0,20)}
+            {deputadosAExibir.length !== 0 ? deputadosAExibir.slice(0,20) : <div></div>}
           </FlipMove>
         </Card>
       </div>
@@ -91,6 +127,76 @@ class DeputadosContainer extends Component {
 
     return nVotacoesDepObject;
   }
+
+  buscaNome(evento){
+    var nomeDeputado = evento.target.value;
+    this.setState({filterName: nomeDeputado})
+    var deputadosNome = this.filtraPorNome(nomeDeputado);
+    this.setState({deputadosPorNome: deputadosNome})
+  }
+
+  buscaUF(evento){
+    var UFDeputado = evento.target.value;
+    this.setState({filterUf: UFDeputado})
+
+    console.log(UFDeputado);
+    var deputados = this.filtraPorUf(UFDeputado);
+
+    this.setState({deputadosPorUf: deputados})
+  }
+
+  buscaPartido(evento){
+    var partidoDeputado = evento.target.value;
+    this.setState({filterPartido: partidoDeputado})
+
+    console.log(partidoDeputado);
+    var deputados = this.filtraPorPartido(partidoDeputado);
+
+    this.setState({deputadosPorPartido: deputados})
+  }
+
+  filtraPorPartido(partidoDeputado){
+    let deputados = DeputadoFactory.inicializaComponentesDeputados("mobile",this.props.scoreDeputados, this.nVotacoesDep, this.props.votosSimilares, this.props.nVotosUsuario);
+    let deputadosFiltrados = deputados.filter(d => ((d.props.children.props.partido.toLowerCase()) === partidoDeputado.toLowerCase()));
+    return deputadosFiltrados.map(d => d.key);
+  }
+
+  filtraPorUf(UFDeputado){
+    let deputados = DeputadoFactory.inicializaComponentesDeputados("mobile",this.props.scoreDeputados, this.nVotacoesDep, this.props.votosSimilares, this.props.nVotosUsuario);
+    let deputadosFiltrados = deputados.filter(d => ((d.props.children.props.uf.toLowerCase()) === UFDeputado.toLowerCase()));
+    return deputadosFiltrados.map(d => d.key);
+    //this.setState({deputados: deputados})
+    //console.log(deputados);
+  }
+
+  filtraPorNome(nomeDeputado){
+    let deputados = DeputadoFactory.inicializaComponentesDeputados("mobile",this.props.scoreDeputados, this.nVotacoesDep, this.props.votosSimilares, this.props.nVotosUsuario);
+    let deputadosFiltrados = deputados.filter(d => (d.props.children.props.nome.toLowerCase().indexOf(nomeDeputado.toLowerCase()) >= 0));
+    return deputadosFiltrados.map(d => d.key);
+    //this.setState({deputados: deputados})
+  }
+
+  estados(){
+    var estados = ["TODOS","AC", "AL", "AM", "AP", "BA", "CE", "DF", "ES", "GO", "MA",
+                   "MT", "MS", "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN",
+                   "RO", "RS", "RR", "SC", "SE", "SP", "TO"];
+
+    return estados.map(function(a){return <MenuItem key={a} value={a}>{a}</MenuItem>})
+  }
+
+  partidos(){
+    let deputados = DeputadoFactory.inicializaComponentesDeputados("mobile",this.props.scoreDeputados, this.nVotacoesDep, this.props.votosSimilares, this.props.nVotosUsuario);
+    let partidos = new Set();
+
+    partidos.add("TODOS");
+
+    for(var i = 0; i < deputados.length; i++){
+      partidos.add(deputados[i].props.children.props.partido);
+    }
+
+    return Array.from(partidos).map(function(a){return <MenuItem key={a} value={a}>{a}</MenuItem>})
+  }
+
 }
 
 export default DeputadosContainer;
